+++ docs/history.md
@@ -169,6 +169,8 @@ This is the first release candidate for Hatch v1, a complete rewrite.
 
 ### Unreleased
 
+### [0.23.0](https://github.com/ofek/hatch/releases/tag/hatchling-v0.23.0) - 2022-04-28 ### {: #hatchling-v0.23.0 }
+
 ***Added:***
 
 - Add a `version` command